requirements.txt
fastapi
uvicorn
google-generativeai
pytest
httpx
sentence-transformers
faiss-cpu